refactor to smaller files, remove inline styles
